Fix the application of dirichlet BCs
The application of the dirichlet BCs is now fixed so that each point is
only set a single time. This is important/useful because it allows us to
also set the displacement deltas when the dirichlet BC is set. This is
critical for use in the biotissue code when incremental deformation
gradients are used. closes biotissue/#143